Update paylaod of /api/cse/system, move some constants around
