+++ container_service_extension/common/constants/server_constants.py
@@ -42,9 +42,6 @@ PKS_COMPUTE_PROFILE_KEY = 'pks_compute_profile_name'
 # PKS API endpoint version
 VERSION_V1 = 'v1'
 
-# vCD API versions supported by CSE
-SUPPORTED_VCD_API_VERSIONS = ['33.0', '34.0', '35.0', '36.0']
-
 # CSE global pvdc compute policy name
 CSE_GLOBAL_PVDC_COMPUTE_POLICY_NAME = 'global'
 CSE_GLOBAL_PVDC_COMPUTE_POLICY_DESCRIPTION = \
